backend/requirements.txt
Flask
flask_cors
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
openai
numpy
python-dotenv
